backend/requirements.txt
Django==5.1.1
django-cors-headers==4.0.0
gunicorn==20.1.0
environs==9.5.0
django-ninja==1.3.0
Pillow==9.5.0
psycopg2-binary==2.9.6
psycopg2==2.9.6
pytz~=2023.3.post1
python-slugify==8.0.1
requests==2.31.0